+++ test/cppgrep_help_test.cpp
@@ -7,5 +7,5 @@
 int main()
 {
     std::array<const char*, 2> argv { "cppgrep", "-h" };
-    return klang::cppgrep::main(argv.size(), argv.data());
+    return static_cast<int>(klang::cppgrep::main(argv.size(), argv.data()));
 }
